refactor: replace UriBuilder with SelectionUriBuilder for improved URI handling
